+++ src/Properties/ModelPropertyExtension.php
@@ -6,27 +6,19 @@ namespace NunoMaduro\Larastan\Properties;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
-use Iterator;
-use PHPStan\Parser\CachedParser;
 use PHPStan\PhpDoc\TypeStringResolver;
 use PHPStan\Reflection\Annotations\AnnotationsPropertiesClassReflectionExtension;
 use PHPStan\Reflection\ClassReflection;
 use PHPStan\Reflection\PropertiesClassReflectionExtension;
 use PHPStan\Reflection\PropertyReflection;
+use PHPStan\ShouldNotHappenException;
 use PHPStan\Type\IntegerType;
-use RecursiveDirectoryIterator;
-use RecursiveIteratorIterator;
-use RegexIterator;
-use SplFileInfo;
 
 /**
  * @internal
  */
 final class ModelPropertyExtension implements PropertiesClassReflectionExtension
 {
-    /** @var CachedParser */
-    private $parser;
-
     /** @var SchemaTable[] */
     private $tables = [];
 
@@ -39,11 +31,14 @@ final class ModelPropertyExtension implements PropertiesClassReflectionExtension
     /** @var AnnotationsPropertiesClassReflectionExtension */
     private $annotationExtension;
 
-    public function __construct(CachedParser $parser, TypeStringResolver $stringResolver, AnnotationsPropertiesClassReflectionExtension $annotationExtension)
+    /** @var MigrationHelper */
+    private $migrationHelper;
+
+    public function __construct(TypeStringResolver $stringResolver, AnnotationsPropertiesClassReflectionExtension $annotationExtension, MigrationHelper $migrationHelper)
     {
-        $this->parser = $parser;
         $this->stringResolver = $stringResolver;
         $this->annotationExtension = $annotationExtension;
+        $this->migrationHelper = $migrationHelper;
     }
 
     public function hasProperty(ClassReflection $classReflection, string $propertyName): bool
@@ -65,7 +60,7 @@ final class ModelPropertyExtension implements PropertiesClassReflectionExtension
         }
 
         if (count($this->tables) === 0) {
-            $this->initializeTables();
+            $this->tables = $this->migrationHelper->initializeTables();
         }
 
         if ($propertyName === 'id') {
@@ -73,9 +68,17 @@ final class ModelPropertyExtension implements PropertiesClassReflectionExtension
         }
 
         $modelName = $classReflection->getNativeReflection()->getName();
-        /** @var Model $modelInstance */
-        $modelInstance = new $modelName;
-        $tableName = $modelInstance->getTable();
+
+        try {
+            $reflect = new \ReflectionClass($modelName);
+
+            /** @var Model $modelInstance */
+            $modelInstance = $reflect->newInstanceWithoutConstructor();
+
+            $tableName = $modelInstance->getTable();
+        } catch (\ReflectionException $e) {
+            return false;
+        }
 
         if (! array_key_exists($tableName, $this->tables)) {
             return false;
@@ -105,9 +108,18 @@ final class ModelPropertyExtension implements PropertiesClassReflectionExtension
     ): PropertyReflection {
         $modelName = $classReflection->getNativeReflection()->getName();
 
-        /** @var Model $modelInstance */
-        $modelInstance = new $modelName;
-        $tableName = $modelInstance->getTable();
+        try {
+            $reflect = new \ReflectionClass($modelName);
+
+            /** @var Model $modelInstance */
+            $modelInstance = $reflect->newInstanceWithoutConstructor();
+
+            $tableName = $modelInstance->getTable();
+        } catch (\ReflectionException $e) {
+            // `hasProperty` should return false if there was a reflection exception.
+            // so this should never happen
+            throw new ShouldNotHappenException();
+        }
 
         if (
             (! array_key_exists($tableName, $this->tables)
@@ -129,46 +141,6 @@ final class ModelPropertyExtension implements PropertiesClassReflectionExtension
             $this->stringResolver->resolve($column->readableType),
             $this->stringResolver->resolve($column->writeableType)
         );
-    }
-
-    private function initializeTables(): void
-    {
-        if (! is_dir(database_path().'/migrations')) {
-            return;
-        }
-
-        $schemaAggregator = new SchemaAggregator();
-        $files = $this->getMigrationFiles(database_path().'/migrations');
-        $filesArray = iterator_to_array($files);
-        ksort($filesArray);
-
-        $this->requireFiles($filesArray);
-
-        foreach ($filesArray as $file) {
-            $schemaAggregator->addStatements($this->parser->parseFile($file->getPathname()));
-        }
-
-        $this->tables = $schemaAggregator->tables;
-    }
-
-    /**
-     * @param string $path
-     *
-     * @return Iterator<SplFileInfo>
-     */
-    private function getMigrationFiles(string $path): Iterator
-    {
-        return new RegexIterator(new RecursiveIteratorIterator(new RecursiveDirectoryIterator($path)), '/\.php$/i');
-    }
-
-    /**
-     * @param SplFileInfo[] $files
-     */
-    private function requireFiles(array $files): void
-    {
-        foreach ($files as $file) {
-            require_once $file;
-        }
     }
 
     private function getDateClass(): string